Remove dependence on PngWriter
Now we don't need to link it into tool binaries!
Also removed unused MsgTimer functionality
